requirements.txt
streamlit
requests
python-dotenv
streamlit-player
fastapi
uvicorn[standard]
chromadb
sentence-transformers
mcp
starlette
pypdf
python-multipart
